[FE] refactor: 로딩 상태 및 css 구조 개선

See merge request s12-final/S12P31C204!186

+++ frontend/roll-model/src/app/workspace/data-preprocess/complete/page.tsx
@@ -23,7 +23,7 @@ const categoryMap: Record<string, { label: string; description: string }> = {
 
 const columnTypeMap: Record<string, string> = {
   integer: '정수',
-  float: '소수',
+  double: '소수',
   string: '문자열',
   boolean: '참/거짓',
   datetime: '날짜/시간',
